+++ spec/factories/task.rb
@@ -2,7 +2,7 @@ FactoryBot.define do
   factory :task do
     title { Faker::Lorem.sentence }
     content { Faker::Lorem.sentences }
-    state { Faker::Lorem.word }
+    state { Task::STATES.sample }
 
     project
   end